+++ rest/src/interfaces/middlewares/catchHttpClientError/catchHttpClientError.interface.ts
@@ -1,0 +1,37 @@
+import {CatchHttpClientErrorBehavior, HttpClientErrorHandlers, HttpClientErrorMessages} from '@anglr/error-handling';
+
+/**
+ * Configuration that can be used within catch http client error middleware
+ */
+export interface RestCatchHttpClientError extends TypedPropertyDescriptor<unknown>
+{
+    /**
+     * Behavior of catch http client error, default is `CatchHttpClientErrorBehavior.Suppress`
+     */
+    behavior: CatchHttpClientErrorBehavior|undefined|null;
+
+    /**
+     * Indication whether skip displaying of notifications for errors
+     */
+    skipErrorNotifications: boolean|undefined|null;
+
+    /**
+     * Indication whether skip server validation errors processing
+     */
+    skipServerValidationErrors: boolean|undefined|null;
+
+    /**
+     * Indication whether forcibly display custom message if present, not only as fallback but always
+     */
+    forceCustomMessageDisplay: boolean|undefined|null;
+
+    /**
+     * Object storing default messages to be displayed for specific http status codes
+     */
+    messages: HttpClientErrorMessages|undefined|null;
+
+    /**
+     * Object storing default error handlers for specific http status codes
+     */
+    handlers: HttpClientErrorHandlers|undefined|null;
+}