+++ rest/src/interfaces/middlewares/httpClientErrorProcessing/httpClientErrorProcessing.interface.ts
@@ -1,7 +1,7 @@
 import {HttpClientErrorResponseMapper, HttpClientValidationErrorResponseMapper} from '@anglr/error-handling';
 
 /**
- * Contains custom mapper functions
+ * Configuration that can be used within http client error processing middleware
  */
 export interface RestHttpClientErrorProcessing extends TypedPropertyDescriptor<unknown>
 {
@@ -14,4 +14,9 @@ export interface RestHttpClientErrorProcessing extends TypedPropertyDescriptor<u
      * Response mapper for http client validation errors
      */
     clientValidationErrorsResponseMapper: HttpClientValidationErrorResponseMapper|undefined|null;
+
+    /**
+     * Array of ignored http status codes that will not be processed
+     */
+    ignoredHttpStatusCodes: number[]|undefined|null;
 }