+++ rest/src/interfaces/middlewares/index.ts
@@ -1,1 +1,2 @@
+export * from './catchHttpClientError/catchHttpClientError.interface';
 export * from './httpClientErrorProcessing/httpClientErrorProcessing.interface';