+++ rest/src/middlewares/index.ts
@@ -1,2 +1,3 @@
+export * from './catchHttpClientError.middleware';
 export * from './clientErrorHandling.middleware';
 export * from './httpClientErrorProcessing.middleware';