+++ src/rxjsOperators/catchHttpClientError.operator.ts
@@ -1,9 +1,12 @@
 import {Injector, inject} from '@angular/core';
+import {extend} from '@jscrpt/common';
 import {Observable, OperatorFunction, catchError, throwError} from 'rxjs';
 
 import {CatchHttpClientErrorOptions} from '../interfaces';
 import {handleHttpClientErrors} from '../errorHandlers';
 import {HttpClientError} from '../misc/classes/httpClientError';
+import {HttpClientErrorHandlers, HttpClientErrorMessages} from '../misc/types';
+import {HTTP_CLIENT_ERROR_HANDLERS, HTTP_CLIENT_ERROR_MESSAGES} from '../misc/tokens';
 
 /**
  * Catches http client errors and handles them according provided options
@@ -12,6 +15,8 @@ import {HttpClientError} from '../misc/classes/httpClientError';
 export function catchHttpClientError<TIn>(options?: CatchHttpClientErrorOptions): OperatorFunction<TIn, TIn|HttpClientError>
 {
     const injector: Injector = options?.injector ?? inject(Injector);
+    const messages: HttpClientErrorMessages = extend({}, injector.get(HTTP_CLIENT_ERROR_MESSAGES, {}, {optional: true}), options?.messages);
+    const handlers: HttpClientErrorHandlers = extend({}, injector.get(HTTP_CLIENT_ERROR_HANDLERS, {}, {optional: true}), options?.handlers);
 
     return source =>
     {
@@ -33,6 +38,8 @@ export function catchHttpClientError<TIn>(options?: CatchHttpClientErrorOptions)
                         {
                             ...options,
                             injector,
+                            messages,
+                            handlers,
                         });
     
                         if(!result)